# contracts/openapi/operations.v1.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Cutover operations API",
    "version": "1.0.0",
    "description": "Local synthetic operations. Every request checks role and site. Errors use RFC 9457: 400 malformed, 401 unauthenticated, 403 role denied, 404 outside site or absent, 409 conflict, 413 body over 64 KiB, 422 domain validation, 429 rate limit, 503 durable admission unavailable with Retry-After."
  },
  "servers": [
    {
      "url": "http://localhost:8780"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
    "/api/v1/sites/{siteId}/orders": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        }
      ],
      "post": {
        "operationId": "createOrders",
        "summary": "Durably accept synthetic work; acceptance is not completion",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/order-request"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Committed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/accepted"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      },
      "get": {
        "operationId": "listOrders",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reference",
            "in": "query",
            "description": "Case-insensitive full or partial external reference, within the authenticated site.",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "shortagesOnly",
            "in": "query",
            "description": "Include only orders with at least one unreserved requested unit.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current owner projection, with observation time where applicable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order-page"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        },
        "description": "Newest recorded orders first, with a stable timestamp and ID cursor. A cursor outside this site yields an empty page."
      }
    },
    "/api/v1/sites/{siteId}/orders/{id}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getOrders",
        "responses": {
          "200": {
            "description": "Current owner projection, with observation time where applicable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/tasks": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "listTasks",
        "responses": {
          "200": {
            "description": "Current owner projection, with observation time where applicable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/task-list"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/orders/{id}/cancellation": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "cancelOrder",
        "summary": "Supervisor cancels an entirely unstarted order",
        "description": "expectedVersion is the order version. The durable intent holds dispatch while the adapter fences the complete movement inventory. A 200 response confirms exactly one reservation release; 409 indicates a version or unsafe-movement conflict. A 503 CANCELLATION_PENDING means the intent remains durable; inspect the order. Reuse the identical request/key after a transport interruption.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reconciliation-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cancellation committed, with complete adapter fence certificate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cancellationId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "orderId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "state": {
                      "const": "CANCELLED"
                    },
                    "version": {
                      "type": "integer"
                    },
                    "certificate": {
                      "$ref": "#/components/schemas/cancellation-certificate"
                    }
                  },
                  "required": [
                    "cancellationId",
                    "orderId",
                    "state",
                    "version",
                    "certificate"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/problem"
          },
          "401": {
            "$ref": "#/components/responses/problem"
          },
          "403": {
            "$ref": "#/components/responses/problem"
          },
          "404": {
            "$ref": "#/components/responses/problem"
          },
          "409": {
            "$ref": "#/components/responses/problem"
          },
          "413": {
            "$ref": "#/components/responses/problem"
          },
          "422": {
            "$ref": "#/components/responses/problem"
          },
          "503": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
    "/api/v1/sites/{siteId}/orders/{id}/cancellations/{cancellationId}/retry": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "cancellationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "resumeCancellation",
        "summary": "Supervisor resumes a paused cancellation after repair",
        "description": "expectedVersion is the cancellation version. Only PAUSED may resume. The 200 response records a new bounded retry budget for the same intent; poll the order for the final outcome. The original browser request key is not required.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reconciliation-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Retry recorded, with cancellationId, orderId, state PENDING and version"
          },
          "400": {
            "$ref": "#/components/responses/problem"
          },
          "401": {
            "$ref": "#/components/responses/problem"
          },
          "403": {
            "$ref": "#/components/responses/problem"
          },
          "404": {
            "$ref": "#/components/responses/problem"
          },
          "409": {
            "$ref": "#/components/responses/problem"
          },
          "413": {
            "$ref": "#/components/responses/problem"
          },
          "422": {
            "$ref": "#/components/responses/problem"
          },
          "503": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/tasks/{id}/recovery": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "recoverLegacyTask",
        "description": "Supervisor resumes six exhausted transport observations with expected version and reason. Retains movement/allocation/command identities.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/reconciliation-request.v1.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audited status investigation requested"
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
    "/api/v1/sites/{siteId}/tasks/{id}": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "maxLength": 40
          }
        },
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Read one task from its owning service, scoped to the authenticated site.",
        "responses": {
          "200": {
            "description": "Current owner task projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/task-view"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/problem"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "order-request": {
        "$ref": "../schemas/order-request.v1.json"
      },
...
      "reconciliation-request": {
        "$ref": "../schemas/reconciliation-request.v1.json"
      },
      "problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "accepted": {
        "type": "object",
        "required": [
          "id",
          "statusUrl"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "statusUrl": {
            "type": "string"
          }
        }
      },
      "order-view": {
        "$ref": "../schemas/order-view.v1.json"
      },
      "order-page": {
        "$ref": "../schemas/order-page.v1.json"
      },
      "task-list": {
        "$ref": "../schemas/task-list.v1.json"
      },
...
      "cancellation-certificate": {
        "$ref": "../schemas/cancellation-certificate.v1.json"
      },
...
      "task-view": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "movementId": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneId": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "epoch": {
            "type": [
              "integer",
              "null"
            ]
          },
          "version": {
            "type": "integer"
          },
          "lastError": {
            "type": [
              "string",
              "null"
            ]
          },
          "eligibleAt": {
            "type": "string",
            "format": "date-time"
          },
          "transportFailures": {
            "type": "integer",
            "minimum": 0
          },
          "transportPaused": {
            "type": "boolean"
          },
          "dispatchAcceptedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "movementId",
          "orderId",
          "zoneId",
          "state",
          "owner",
          "epoch",
          "version",
          "lastError",
          "eligibleAt"
        ],
        "additionalProperties": true
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "problem": {
        "description": "Bounded problem response without internal exceptions",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/problem"
            }
          }
        }
      }
    }
  }
}